refactor: divide tests into sub-directories
