+++ .mocharc.json
@@ -1,0 +1,4 @@
+{
+    "recursive": true,
+    "require": "test/helpers.js"
+}
